pyproject.toml
[project]
name = "stt-diarization-api"
version = "0.1.0"
description = "Speech-to-Text API with Speaker Diarization"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn",
  "python-multipart",
  "openai-whisper",
  "torch",
  "torchaudio",
  "celery",
  "redis",
  "pyannote.audio",
  "python-dotenv",
  "soundfile",
  "numpy",
]